+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/PlaceholderDecorator` to render the wrapped component when it comes into view
+- `ui/PlaceholderDecorator`, `ui/PlaceholderContainer` to render the wrapped component when it comes into view
 
 ## [1.0.0-beta.4] - 2017-03-10